move popup api hooks a level up
