+++ src/popup/warnings/edit-page.tsx
@@ -1,5 +1,5 @@
 import Layout from '../layout'
-import { useWarning, useUpdateWarningMutation } from './api-hooks'
+import { useWarning, useUpdateWarningMutation } from '../api-hooks'
 import WarningForm from './form'
 import { useParams, useNavigate } from 'react-router'
 import { type Warning } from '../../api'
