+++ src/popup/warnings/list-page.tsx
@@ -4,7 +4,7 @@ import Icon from '../icon'
 import EditIcon from '@fortawesome/fontawesome-free/svgs/solid/pen-to-square.svg'
 import TrashIcon from '@fortawesome/fontawesome-free/svgs/solid/trash.svg'
 import Layout from '../layout'
-import { useAllWarnings, useRemoveWarningMutation } from './api-hooks'
+import { useAllWarnings, useRemoveWarningMutation } from '../api-hooks'
 import { Link } from 'react-router-dom'
 import { type WarningWithId } from '../../api'
 import { type CSSProperties } from 'react'
